Bugzilla BZ#691867: add ldaps support through perLDAP
git-svn-id: svn+ssh://svn.fedorahosted.org/svn/pki/trunk@1938 c9f7a03b-bd48-0410-a16d-cbbf54688b0b

+++ pki/base/tps/lib/perl/PKI/TPS/AdminPanel.pm
@@ -27,6 +27,7 @@ use PKI::TPS::GlobalVar;
 use PKI::TPS::Common;
 use URI::URL;
 use URI::Escape;
+use Mozilla::LDAP::Conn;
 
 package PKI::TPS::AdminPanel;
 $PKI::TPS::AdminPanel::VERSION = '1.00';
@@ -113,6 +114,8 @@ sub update
     my $nickname = $::config->get("preop.cert.sslserver.nickname");
     my $instanceID = $::config->get("service.instanceID");
     my $instanceDir = $::config->get("service.instanceDir");
+    my $certdir = "$instanceDir/alias";
+
     my $db_password = `grep \"internal:\" \"$instanceDir/conf/password.conf\" | cut -c10-`;
     $db_password =~ s/\n$//g;
 
@@ -155,8 +158,9 @@ sub update
     my $admincert = $response->{Requests}->{Request}->{b64};
     &PKI::TPS::Wizard::debug_log("AdminPanel: admincert " . $admincert);
 
-    my $ldap_host = $::config->get("preop.database.host");
-    my $ldap_port = $::config->get("preop.database.port");
+    my $hostport = $::config->get("auth.instance.1.hostport");
+    my ($ldap_host, $ldap_port) = split(/:/, $hostport);
+    my $secureconn = $::config->get("auth.instance.1.ssl");
     my $basedn = $::config->get("preop.database.basedn");
     my $binddn = $::config->get("preop.database.binddn");
 #    my $bindpwd = $::config->get("tokendb.bindPass");
@@ -168,16 +172,31 @@ sub update
     my $flavor = "pki";
     $flavor =~ s/\n//g;
 
-    my $ldapmodify_path = "/usr/bin/ldapmodify";
+    my $conn =  PKI::TPS::Common::make_connection(
+                  {host => $ldap_host, port => $ldap_port, pswd => $bindpwd, bind => $binddn, cert => $certdir},
+                  $secureconn);
 
+    if (!$conn) {
+      &PKI::TPS::Wizard::debug_log("AdminPanel: Failed to connect to the internal database");
+      $::symbol{errorString} = "Failed to connect to the internal database";
+      return 0;
+    };
+
+    my $msg;
     $admincert =~ s/\//\\\//g;
     system("sed -e 's/\$TOKENDB_ROOT/$basedn/' " .
               "-e 's/\$TOKENDB_AGENT_PWD/$password/' " .
               "-e 's/\$TOKENDB_AGENT_CERT/$admincert/' " .
               "/usr/share/$flavor/tps/scripts/addAgents.ldif > $tmp");
-    system("$ldapmodify_path -x -h '$ldap_host' -p '$ldap_port' -D '$binddn' " .
-              "-w '$bindpwd' -a " .
-              "-f '$tmp'");
+    if (! &PKI::TPS::Common::import_ldif($conn, $tmp, \$msg)) {
+      &PKI::TPS::Wizard::debug_log("AdminPanel: $msg");
+      $::symbol{errorString} = "Failed to add agents to database";
+      $conn->close();
+      return 0;
+    };
+    if ($msg ne "") {
+      &PKI::TPS::Wizard::debug_log("AdminPanel: adding agents errors : $msg");
+    }
     system("rm $tmp");
 
     my $reqid = $response->{Requests}->{Request}->{Id};
